requirements.txt
flake8==6.0.0
numpy==1.24.2
pytest==7.3.0
aiogram==2.25.1
python-dotenv==1.0.0
requests==2.28.2
Django==4.2
django-taggit==3.1.0
Markdown==3.4.3
ipython==8.13.1
factory-boy==3.2.1
coverage==7.2.5
